Move all filters to filters file
