+++ client/src/utils/filters.js
@@ -1,6 +1,9 @@
 /* eslint-disable import/prefer-default-export */
 
+import fuzzysort from 'fuzzysort';
+
 import { DAYS_OF_WEEK } from 'data/resources';
+
 
 const dayMs = 1000 * 60 * 60 * 24;
 // Get the days of the week that appear in a given date range
@@ -28,3 +31,52 @@ export function matchesDays(resource, daysOfWeek) {
   const matchDaysOfWeek = dowsInRange(resource.startDate, resource.endDate);
   return daysOfWeek.some((d) => matchDaysOfWeek.includes(d));
 }
+
+
+export default function filterResources(passedResources, filters) {
+  let resources = passedResources;
+  // Text search
+  if (filters.search) {
+    const results = fuzzysort.go(
+      filters.search,
+      resources,
+      { keys: ['name', 'organization'] },
+    );
+    resources = results.map((result) => result.obj);
+  }
+  // Other filters
+  resources = resources.filter((r) => {
+    let include = true;
+    // Category filter
+    if (filters.categories?.length) { include &&= filters.categories.includes(r.category); }
+    // Day of week filter
+    if (filters.daysOfWeek?.length) { include &&= matchesDays(r, filters.daysOfWeek); }
+    // Date filter
+    if (filters.startMonth && filters.startDay && filters.startYear) {
+      include &&= r.endDate && new Date(r.endDate) >= new Date(`${filters.startYear}-${filters.startMonth}-${filters.startDay}`);
+    }
+    if (filters.endMonth && filters.endDay && filters.endYear) {
+      include &&= r.startDate && new Date(r.startDate) <= new Date(`${filters.endYear}-${filters.endMonth}-${filters.endDay}`);
+    }
+
+    return include;
+  });
+
+  return resources;
+}
+
+
+// Latitude/longitude filters happen in a separate step so that the map can display everything
+
+export const geoFilterResources = (
+  resources,
+  { min_lat: minLat, max_lat: maxLat, min_lng: minLng, max_lng: maxLng },
+) => resources.filter(({ location }) => {
+  if ((minLat || maxLat || minLng || maxLng) && !location) return false;
+  const { latitude: resourceLat, longitude: resourceLng } = location || {};
+  if (minLat && (resourceLat < minLat)) return false;
+  if (maxLat && (resourceLat > maxLat)) return false;
+  if (minLng && (resourceLng < minLng)) return false;
+  if (maxLng && (resourceLng > maxLng)) return false;
+  return true;
+});
